+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 * Add `u.if` to conditionally update objects.
 * Add `u.map` to update all values in an array or object.
+* Replace object outright if null or constant givens as `updates`.
 
 ## [0.3.1]
 * Actually expose `u.in`.